Establish basic functionality of the game
I built the necessary functions to create players,
create a game board and to play the game. I also
started to build a function to check for a winner.

So far the game can only be played via the console
and currently no winner is correctly detected, yet.
The isWinner function does work when evaluating columns.