Further work on the display functions
I implemented a function that retrieves the coordinates
from the clicking on the displayed game board. It's
also checking wether the cell is still available.
Next up I need to get the game working so that clicking
on an empty cell will switch players, etc.
Also the createPlayers function is causing troubles
when trying to set the playerName via the input field.
The cause of the issue is that currently the entire
createPlayer function is called whenever one accesses
any of its inherit functions and thereby overwriting
the previous array of players.